Address Copilot PR review: fix infinite redirect loop for non-admin sessions
proxy.ts redirected any authenticated Supabase user away from /login --
but "has a Supabase session" isn't the same as "is an active admin" (a
suspended admin, or an orphaned Auth user with no admin_users row, has
the former but not the latter). For that user: (protected)/layout.tsx
sends them to /login because getAdminSession() correctly finds no active
admin row, then proxy.ts immediately bounced them back to / because it
only checked for a session, forever.

Fix: proxy.ts no longer redirects away from /login at all. /login is now
a server component that does the same getAdminSession() check as every
protected page before deciding whether to redirect or render the form
(moved into LoginForm.tsx, a client component).
