+++ admin/proxy.ts
@@ -3,13 +3,19 @@ import { NextResponse, type NextRequest } from "next/server";
 
 // Standard Supabase SSR proxy (Next.js 16 renamed "middleware" to "proxy" --
 // same file conventions otherwise): refreshes the session cookie on every
-// request (required so sessions don't silently expire) and does a coarse
-// redirect based on whether a session exists at all. This is NOT the real
-// authorization check -- it can't be, it doesn't know about admin_users
-// roles/status. The real check is getAdminSession() in
-// app/(protected)/layout.tsx, which runs after this on every protected page.
-// Both layers matter: this one is what keeps someone who was suspended
-// mid-session from riding out a stale cookie past the login redirect.
+// request (required so sessions don't silently expire) and redirects
+// requests with no Supabase session at all away from protected routes.
+//
+// Deliberately does NOT redirect an authenticated user away from /login --
+// "has a Supabase session" is not the same as "is an active admin" (a
+// suspended admin, or an orphaned Auth user with no admin_users row, has
+// the former but not the latter), and this layer has no cheap way to check
+// the latter without a DB round trip on every request. /login does that
+// real check itself via getAdminSession() before rendering the form, same
+// as app/(protected)/layout.tsx does for every protected page. Redirecting
+// from here on "has a session" alone previously caused an infinite loop:
+// (protected)/layout.tsx sends a non-admin session to /login, this proxy
+// immediately bounced them back to /, forever.
 export async function proxy(request: NextRequest) {
   let supabaseResponse = NextResponse.next({ request });
 
@@ -45,12 +51,6 @@ export async function proxy(request: NextRequest) {
   if (!user && !isAuthRoute) {
     const url = request.nextUrl.clone();
     url.pathname = "/login";
-    return NextResponse.redirect(url);
-  }
-
-  if (user && request.nextUrl.pathname === "/login") {
-    const url = request.nextUrl.clone();
-    url.pathname = "/";
     return NextResponse.redirect(url);
   }
